requirements.txt
openai
openai-agents
fastmcp
fastapi
uvicorn
python-dotenv
authlib
itsdangerous
oci
